requirements.txt
django-heroku
whitenoise